requirements.txt
altair==6.0.0
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.63
botocore==1.42.63
cachetools==7.0.5
captcha==0.7.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.2
contourpy==1.3.3
cryptography==46.0.6
curl_cffi==0.13.0
cycler==0.12.1
extra-streamlit-components==0.1.81
filelock==3.24.3
fonttools==4.61.1
frozendict==2.4.7
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
multitasking==0.0.12
narwhals==2.18.1
numpy==2.4.2
packaging==26.0
pandas==3.0.1
peewee==4.0.4
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
protobuf==7.34.1
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
PyJWT==2.12.1
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
seaborn==0.13.2
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
streamlit==1.56.0
streamlit-authenticator==0.4.2
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
watchdog==6.0.0
websockets==16.0
yfinance==1.2.0
